refactor: move prepare-swagger-ui to util
