+++ package.json
@@ -4,7 +4,7 @@
   "description": "Generate Swagger files automatically for Fastify.",
   "main": "index.js",
   "scripts": {
-    "prepare": "node prepare-swagger-ui",
+    "prepare": "node lib/util/prepare-swagger-ui",
     "test": "standard && tap --100 test/*.js test/esm/*.js && npm run typescript",
     "prepublishOnly": "npm run prepare",
     "typescript": "tsd"
